This commit will implement multiple changes: Task 15... 1. Will respond with 200 when sorting multiple columns by ASC and DESC 2. throw an error when invalid params are entered 3. will test for sql injection by actually inputting sql injection and throw an error when it happens 4. add router files for each part of the api 5. add controllers and models for each part of the api depending on its function 6. refactor multiple files 7. remove checkingdata.js
