--- checkingdata.js
@@ -1,6 +1,0 @@
-// const data = require('./db/data/test-data/index');
-
-// console.log(data.userData);
-// console.log(data.commentData);
-// console.log(data.topicData);
-// console.log(data.articleData);
