+++ db/connection.js
@@ -1,10 +1,6 @@
 const { Pool } = require('pg');
 const ENV = process.env.NODE_ENV || 'development';
 
-// require('dotenv').config({
-//   path: `${__dirname}/../.env.${ENV}`,
-// });
-
 const pathToEnvFile = `${__dirname}/../.env.${ENV}`
 
 require('dotenv').config({path: pathToEnvFile})
