+++ endpoints.json
@@ -10,41 +10,34 @@
     }
   },
   "GET /api/articles": {
-    "description": "serves an array of all articles",
-    "queries": ["author", "topic", "sort_by", "order"],
-    "exampleResponse": {
-      "articles": [
-        {
-          "title": "UNCOVERED: catspiracy to bring down democracy",
-          "topic": "cats",
-          "author": "rogersop",
-          "created_at": "2020-08-03T13:14:00.000Z",
-          "article_img_url": "https://images.pexels.com/photos/158651/news-newsletter-newspaper-information-158651.jpeg?w=700&h=700",
-          "votes": 0,
-          "article_id": 5,
-          "comment_count": 2
-        }
-      ]
+    "description": "serves an array of all articles, can have sort_by and order queries",
+    "queries": {
+      "sort_by": [
+        "created_at",
+        "votes",
+        "comment_count",
+        "title",
+        "topic",
+        "author"
+      ],
+      "order": ["ASC", "DESC"],
+      "exampleResponse": {
+        "articles": [
+          {
+            "title": "UNCOVERED: catspiracy to bring down democracy",
+            "topic": "cats",
+            "author": "rogersop",
+            "created_at": "2020-08-03T13:14:00.000Z",
+            "article_img_url": "https://images.pexels.com/photos/158651/news-newsletter-newspaper-information-158651.jpeg?w=700&h=700",
+            "votes": 0,
+            "article_id": 5,
+            "comment_count": 2
+          }
+        ]
+      }
     }
   },
-  "GET /api/articles (topic query)": {
-    "description": "serves an array of all articles with specified query topic",
-    "queries": [ "sort_by", "order", "topic"],
-    "exampleResponse": {
-      "articles": [
-        {
-          "title": "UNCOVERED: catspiracy to bring down democracy",
-          "topic": "cats",
-          "author": "rogersop",
-          "created_at": "2020-08-03T13:14:00.000Z",
-          "article_img_url": "https://images.pexels.com/photos/158651/news-newsletter-newspaper-information-158651.jpeg?w=700&h=700",
-          "votes": 0,
-          "article_id": 5,
-          "comment_count": 2
-        }
-      ]
-    }
-  },
+
   "GET /api/articles/:article_id": {
     "description": "fetches an object of inputted article_id",
     "queries": [],
@@ -105,17 +98,19 @@
   "GET /api/users": {
     "description": "serves an array of all users",
     "queries": [],
-    "exampleResponse": { "users": [
-      {
-        "username": "username1",
-        "name": "actual name",
-        "avatar_url": "https://www.iamalinkhello.com"
-      },
-      {
-        "username": "icellusedkars",
-        "name": "sam",
-        "avatar_url": "https://avatars2.githubusercontent.com/u/24604688?s=460&v=4"
-      }
-    ]}
+    "exampleResponse": {
+      "users": [
+        {
+          "username": "username1",
+          "name": "actual name",
+          "avatar_url": "https://www.iamalinkhello.com"
+        },
+        {
+          "username": "icellusedkars",
+          "name": "sam",
+          "avatar_url": "https://avatars2.githubusercontent.com/u/24604688?s=460&v=4"
+        }
+      ]
+    }
   }
 }
