+++ models/index.js
@@ -1,15 +1,8 @@
-const {
-  fetchTopics,
-  fetchArticleById,
-  fetchArticles,
-  fetchArticleCommentsById,
-  insertCommentToArticle,
-  changeVoteAmount,
-  removeCommentById,
-  fetchUsers,
-} = require("./app.models");
-
-const { checkUserExists, checkCommentExists } = require("./checker.models");
+const { fetchTopics } = require("./topics.models");
+const { fetchArticles, fetchArticleById, fetchArticleCommentsById } = require("./articles.models");
+const { fetchUsers, checkUserExists } = require("./users.models");
+const { removeCommentById, insertCommentToArticle, checkCommentExists } = require("./comments.models");
+const { changeVoteAmount } = require("./votes.models");
 
 module.exports = {
   fetchTopics,
