+++ models/topics.models.js
@@ -1,0 +1,13 @@
+const db = require("../db/connection");
+
+exports.fetchTopics = () => {
+    return db
+      .query("SELECT * FROM topics")
+      .then((res) => {
+        return res.rows;
+      })
+      .catch((err) => {
+        next(err);
+      });
+  };
+
