+++ routes/api.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const { getApi } = require('../controllers/index');
+
+const router = express.Router();
+
+router.get('/', getApi);
+
+module.exports = router;
