+++ routes/topics.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const { getTopics } = require('../controllers/index');
+
+const router = express.Router();
+
+router.get('/', getTopics);
+
+module.exports = router;
